+++ pcsd/remote.rb
@@ -964,8 +964,8 @@ def setup_cluster(params, request, auth_user)
   end
   nodes_options = nodes + options
   nodes_options += options_udp if transport_udp
-  if params[:no_hardened] == "1"
-      nodes_options << "--no-hardened"
+  if ['0', '1'].include?(params[:encryption])
+      nodes_options << "--encryption=#{params[:encryption]}"
   end
   stdout, stderr, retval = run_cmd(
     auth_user, PCS, "cluster", "setup", "--enable", "--start", "--async",